Add MDETR text encoder
This commit adds support for MDETR's RoBERTa text encoder.
We make some modifications to TorchText's RoBERTa encoder to make
things more modular and ensure the implementation matches the one from
Hugging Face (as used by MDETR). The changes include

1) A class MDETRTextEmbeddings. This is similar to the TextEmbeddings class from
FLAVA, but handles position IDs differently.

2) A WrappedTransformerEncoder class. This class will take MDETR text embeddings
and feed them through the transformer encoder. The main difference between this class
and the TorchText one is that the TorchText class includes embeddings as part of the encoder.

3) A class MDETRTextEncoder to stitch (1) and (2) together

We add a test for each of these three classes and compare to the result obtained when
loading directly from a pretrained MDETR checkpoint.

Similar to the image backbone commit, once the changes look good at a high level I will
add docstrings and the like.

Tested with
```
$ python -m pytest -v test/modules/encoders/test_mdetr_text_encoder.py
    test/modules/encoders/test_mdetr_text_encoder.py::TestMDETRTextEncoder::test_mdetr_text_embeddings PASSED
    test/modules/encoders/test_mdetr_text_encoder.py::TestMDETRTextEncoder::test_mdetr_text_encoder PASSED
    test/modules/encoders/test_mdetr_text_encoder.py::TestMDETRTextEncoder::test_mdetr_wrapped_transformer PASSED
```

ghstack-source-id: 77a81c5c4a2b07b8a3d30fd9b60cc7ba7b3714e9
Pull Request resolved: https://github.com/facebookresearch/multimodal/pull/62

Changes to text encoder